pyproject.toml
[project]
name = "aidos"
version = "0.1.0"
description = "Айдой — қазақ тілді AI көмекші (дауыс + мәтін)"
requires-python = ">=3.12"
dependencies = [
    "ollama>=0.4.0",
    "openai-whisper>=20231117",
    "edge-tts>=6.1.9",
    "pygame>=2.6.0",
    "requests>=2.32.0",
    "schedule>=1.2.0",
    "rich>=13.7.0",
    "sounddevice>=0.5.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
aidos = "aidos.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aidos"]
